email-archive/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-archive"
version = "2.0.0"
description = "Process email archives (PST files) into a searchable ChromaDB vector database"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "click>=8.1",
    "rich>=13.0",
    "libpff-python>=20231205",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
email-archive = "email_archive.cli:main"

[tool.setuptools.packages.find]
include = ["email_archive*"]

[tool.pytest.ini_options]
testpaths = ["tests_py"]
